moved frame group / frame chunk data classes + parsers to own resp. files
